rename nexon army nft to eralend army nft
