+++ src/strategies/index.ts
@@ -429,7 +429,7 @@ import * as jpegdLockedJpegOf from './jpegd-locked-jpeg-of';
 import * as litDaoGovernance from './lit-dao-governance';
 import * as babywealthyclub from './babywealthyclub';
 import * as battleflyVGFLYAndStakedGFLY from './battlefly-vgfly-and-staked-gfly';
-import * as nexonArmyNFT from './nexon-army-nft';
+import * as eralendArmyNFT from './eralend-army-nft';
 import * as moonbeamFreeBalance from './moonbeam-free-balance';
 import * as stakedotlinkVesting from './stakedotlink-vesting';
 import * as pspInSePSP2Balance from './psp-in-sepsp2-balance';
@@ -880,7 +880,7 @@ const strategies = {
   'lodestar-staked-lp': lodestarStakedLp,
   babywealthyclub,
   'battlefly-vgfly-and-staked-gfly': battleflyVGFLYAndStakedGFLY,
-  'nexon-army-nft': nexonArmyNFT,
+  'eralend-army-nft': eralendArmyNFT,
   'moonbeam-free-balance': moonbeamFreeBalance,
   'stakedotlink-vesting': stakedotlinkVesting,
   'psp-in-sepsp2-balance': pspInSePSP2Balance,
